friendly cube
+ added "com.unity.ai.navigation (ver 1.1.4)" package
+ baked navmesh for entire maze
+ killer cube now follows player around!
