+++ Assets/CubeFollow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bd735a80649213f43bfe448ebc68fbdb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
