+++ Assets/Scenes/Maze.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 764c378eef1f4cd42ac6004e0cae2966
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
