+++ Assets/Scenes/Maze/NavMesh-Maze.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 92274dd4d9c862342b1f7ef324f0279c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
